Make strawberry.scalar work as decorator
Addresses #114.

This change requires the user to call NewType themselves when not using
the decorator form. But this is more consistent in behaviour than
calling it for them. Also, there might be situation where one does not
want to create a new type (e.g. if a specific type already exists).