Add support for custom scalars (#191)
* Implement serialization to custom scalar

Addresses #114.

* Implement deserialization of custom scalars

Addresses #114.

* Allow parse_literal function for custom scalars

Addresses #114.

It is unclear to me in what case this function would be called
and how a test for it could be written.

* Allow to set description of custom scalars

Addresses #114.

* Make strawberry.scalar work as decorator

Addresses #114.

This change requires the user to call NewType themselves when not using
the decorator form. But this is more consistent in behaviour than
calling it for them. Also, there might be situation where one does not
want to create a new type (e.g. if a specific type already exists).

* Add documentation for custom scalars

Addresses #114.

* Add RELEASE.md for custom scalars

Addresses #114.

* Fix code formatting in examples

Co-Authored-By: Patrick Arminio <patrick.arminio@gmail.com>

* Use GraphQL variable instead of format string

* Add custom scalar default serialize/parse_value

* Add Date, DateTime, Time custom scalars

* Do not install pytest-mypy-plugins breaking tests

* Do not recreate default scalar serialize function

* Move DateTime scalars to types/datetime.py

* Make custom scalar tests implementation sync

* Add aniso8601 as optional dependency

for DateTime custom scalars

* Test DateTime custom scalar and add proper names

* Install datetime extras on CI

* Promote aniso8601 to a required dependency

* Ensure parsing literal DateTime values works