Implement `resolve_type` on interfaces
When specifying `is_type_of` we aren't able to return a type
other than the field one, this prevents for example to return
a django model instead of the strawberry type.

So now instead of implementing `is_type_of` on every type that
implements an interface we only implement `resolve_type` on the
interface itself.

See: https://github.com/pythonitalia/pycon/pull/1216#issuecomment-548923736