Remove mention of sequential mutation execution
Whilst correct, this mention doesn't sufficiently explain the behaviour -- namely that mutations themselves are executed sequentially, but the return values are resolved in a second potentially parallel pass (akin to normal query resolution). This is worth discussing in the mutation part of the documentation, but feels out of place in a simple introduction to the basic concepts -- and is potentially confusing.