Fix bug with using arguments with generics
Currently generics work by copying the fields using the
`dataclasses.asdict` method. However this method recursively converts all
the fields of the dataclass into dictionaries. This results in a
TypeError when constructing the schema where the code assumes the
`arguments` attribute is an instance of `ArgumentDefinition` but it's
now a dictionary.

This code fixes the error by replacing the use of `asdict` with
`__dict__` so that it only converts one level deep.