Gracefully handle user-induced subscription errors.
When submitting subscription queries that had syntax errors or invalid
field references the resulting error would cause an unhandled exception
that resulted in the connection being prematurely terminated (without
notice).

In these situations the errors are now handled and a proper `GQL_ERROR`
is returned to the client with the offending error as the payload.