Swap JS CDN for graphqli interface to use unpkg.com
Change is based on a conversation in Discord regarding a number of
reports (including my own) of graphiql-subscriptions-fetcher not loading
from the existing (JSDelivr) CDN.

JKimbo suggested a switch to use unpkg.com which I have confirmed
resolved the issue. Suggested all other files should also be swapped

I have confirmed that all of the files are now loading, whatwg-fetch
required a switch to the un-minified variant with the sha384 hash
updated along with it. The GraphQLI interface is now loading and
operating as expected.

Message linked to discussion thread https://discord.com/channels/689806334337482765/773519351423827978/871285294123581441