Cast custom scalars to the their given type
New versions of pyright doesn't allow to use non-type objects
to annotate dataclasses attributes.

We are casting the return type for now to make pyright think it
we are using the type itself, when in reality we are still using
`ScalarWrapper`.

This should be properly fixed when the new `StrawberryScalar`
is introduced.

Fix #1689