Improve reserved argument injection for resolvers
This commit closes Issue #1666 by improving the injection of reserved
arguments by providing an opt-in functionality to match reserved
arguments by type-annotations. The functionality is added to
`StrawberryResolver`s which is then reused in `StrawberryDirectives`s
via a new `StrawberryDirectiveResolver` class. Additionally, A new
`DirectiveValue` `Annotated` type is defined to allow more flexibility
when defining directive resolvers. To ensure backwards-compatibility, a
fallback to argument name-based matching for reserved parameters is used
for `info` and `value` arguments. Overally, the new `ReservedType`
specification allows arbitrary names for `info` and `value` arguments,
provided that they are annotated as `<my_info_name>: Info` and
`<my_value_name>: DirectiveValue[Any]` respectively.

The new injection functionality is enabled by a
`ReservedParameterSpecification` `Protocol` which is used to define
various types of reserved parameters. Each specification is responsible
for iterating over the `parameters` of the wrapped resolver to `find`
its targeted reserved parameter. This information is then stored in a
mapping of specification to reserved parameter. Later, this mapping is
used to determine which reserved keyword argument(s) to pass to the
resolver during execution.

Summary of Changes:
- Improve reserved argument injection for resolvers by providing an
  opt-in functionality to match reserved arguments by type-annotations.
- Re-use filtering logic of `StrawberryResolver` by defining a
  specialized `StrawberryDirectiveResolver` class
- Support passing `info` and `root` to a `StrawberryDirective`
- Deprecate use of bare `info` and `value` arguments to resolvers
- Allow definition of arbitrarily named keyword arguments to represent
  `info`, and `value` reserved arguments
- Modify @magicmark's failing test to check if a directive without value
  argument correctly recieves its arguments
- Add test to verify new reserved argument injection logic