Update StrawberryField get_result signature to make extending easier
This PR updates the `StrawberryField.get_result` method signature to
only take the field arguments instead of splitting them into python args
and kwargs. This makes extending the field simpler because you no longer
have to consider the `self` param that may or may not be passed to the
underlying resolver.

This PR also adds some initial documentation around implementing
custom fields and updates the code samples for private fields.

Diff-Id: d2b68