Do not use `kw_only=False` when specifying strawberry.field()
The changes in #719 suggest that strawberry dataclasses should always use the `kw_only` mode.
This is great, as it allows mixing the order of fields with/without default values, particularly when inheritance is involved.

However this falls apart when the fields without a default value is specified with `field: type = strawberry.field()`: Due to a minor mistake,
this particular field is now explicitly marked with kw_only=False