Fix strawberry.lazy() when using relative module
When using `Annotated["SomeType", lazy(".some_module")]`, the call stack is inspected to fetch the proper __package__ necessary to resolve `.some_module`.

There was a mistake where theth frame was being inspected inspead of the 3rd.

This wasn't caught in the tests previously because there were 2 frames for `type_a.py` when calling `lazy()`: One for the module definition and one for the class definition:

```
  File "/media/Paulo/Workspace/pass2go/strawberry/tests/schema/test_lazy/type_a.py", line 13, in <module>
    class TypeA:
  File "/media/Paulo/Workspace/pass2go/strawberry/tests/schema/test_lazy/type_a.py", line 19, in TypeA
    def type_b(self) -> Annotated["TypeB", strawberry.lazy(".type_b")]:
  File "/media/Paulo/Workspace/pass2go/strawberry/strawberry/lazy_type.py", line 70, in lazy
    return StrawberryLazyReference(module_path)
  File "/media/Paulo/Workspace/pass2go/strawberry/strawberry/lazy_type.py", line 61, in __init__
    traceback.print_stack()
```

This PR fixes the bug and modifies the test to prevent this issue