Fix dataclass creation in Pydantic integration for Python 3.10.0
Python 3.10.0 missed the `kw_only` parameter for
`dataclasses.make_dataclass()`: https://github.com/python/cpython/issues/89961
It was added in 3.10.1.

This change updates the version checking in the Pydantic integration to
only use the `kw_only` parameter for 3.10.1 versions and above.
Otherwise to falls back to our custom init function.

Diff-Id: 5faf9