Support documenting permissions on the schema
I've been using Strawberry's permission checks and I'm quite satisfied with it,
however the permission information needs to be manually documented, which is error-prone.

This PR adds 2 different ways to expose this in the GraphQL schema: Using schema directives and using descriptions.

The resulting schemas look like this:

```
type Query {
  protectedField: String! @requiresPermissions(permissions: ["MyPermission"])
}
```

or

```
type Query {
  """
  Required permissions:
   - *MyPermission*: Doesn't have permission
  """
  protectedField: String!
}
```

Using directives seem like the best correct solution, but it is likely to be ignored by documentation tools