Fix unresolved arguments using typing.Annotated
Reserved argument parsing, which interally also resolves annotations,
exits early after detecting a reserved argument like an ``info``
argumnent. As a result, any subsequent annotations are not reserved
arguments are unresolved. This results in `StrawberryArgument` not being
able to determine that an argument annotation makes use of
`typing.Annotated`.

This behavior is fixed by by ensuring that `StrawberryArgument` makes
use of the new `StrawberryAnnotation.evaluate` method instead of
consuming the raw annotation from it.

An added benefit of this fix is that by moving annotation resolving code
from `StrawberryResolver` to `StrawberryAnnotation`, it makes it
possible for `StrawberryArgument` and `StrawberryResolver` to share the
same type evaluation cache.

Summary of Changes:
- Move type evaluation and caching logic to `StrawberryAnnotation`
- Separate type evalution from resolution with the new
  `StrawberryAnnotation.evaluate` method. This allows
  `StrawberryArgument` to access the evaluated annotation before it gets
  converted to `StraberryType` specializations
- Remove type evaluation cache from `StrawberryResolver` and move it to
  `StrawberryAnnotation`.
- Fix typing issues in `resolver.py`

Closes #2855