fix: do not strip the type when using auto
`Annotated[auto, ...]` produce issues on python 3.10 and older
due to the fact that it tries to check that the type is `callable`.

This was being masked by the issue that 5bd41027366a04f644feb2679c752cd51b120e08
fixed, due to the fact that extra `Annotated` args were being thrown
away.

This change the evaluation to avoid throwing the original type (usually
`Any`) away, fixing that issue.