refactor: support node_id values containing colons
The node_id can be anything depending on how the user is handling it,
so colons can appear in it, specially when it is some kind of encrypted
value.

This changes how we split the `GlobalID` string to only split the
first colon in it, meaning that the remaining of the string will be
considered as a whole for the node_id.