Handle UNSET values in the strawberry schema. (#3050)
* Handle UNSET values in the strawberry schema.

The codegen currently chokes on `strawberry.UNSET` default values.  This PR updates the codegen to stop raising an exception and instead pass a `GraphQLNullValue(value=UNSET)` to the plugins.  The default python plugin writes a `None` as the default value because there really isn't a well understood concept of UNSET in the python eco
system.  User defined plugins have all the information that they need to handle this differently if they choose to do so.

* Add RELEASE.md file.

---------

Co-authored-by: Matt Gilson <mgilson@lat.ai>