Make relay.BaseNode and provide a nested class for GlobalID
This hacks up the relay.Node and makes a `BaseNode` class that can be
extended by the "public". The `BaseNode` has a new nested `Config` class
with a `GlobalID` attribute that should point to a dataclass that
extends `relay.GlobalID` for customization.

Add in a `scalar_overrides` on the `Schema` definition and it kinda
works?

Example:
All the classes must be named as they are `GlobalID`, `Node`, etc. in order for
all this to work. I'm not sure how I feel about this.

```ruby
@dataclasses.dataclass(order=True, frozen=True)
class GlobalID(relay.GlobalID):
    def __str__(self) -> str:
        return to_gid_url(self.type_name, self.node_id)

    @classmethod
    def from_id(cls, value: Union[str, ID]) -> Self:
        try:
            type_name, node_id = from_gid_url(value)
        except ValueError as e:
            raise relay.GlobalIDValueError(str(e)) from e

        return cls(type_name=type_name, node_id=node_id)

    @classmethod
    def _get_scalar_definition(cls) -> ScalarDefinition:
        return scalar(
            cls,
            name="GlobalID",
            description=GraphQLID.description,
            parse_literal=lambda v, vars=None: cls.from_id(
                GraphQLID.parse_literal(v, vars)
            ),
            parse_value=cls.from_id,
            serialize=str,
            specified_by_url=("https://relay.dev/graphql/objectidentification.htm"),
        )._scalar_definition

@interface(description="An object with a Globally Unique ID")
class Node(relay.BaseNode):
    class Config:
        GlobalID: ClassVar[Type[GlobalID]] = GlobalID

    @field(name="id", description="The Globally Unique ID of this object")
    @classmethod
    def _id(cls, root: Node, info: Info) -> GlobalID:
        return super()._id(root, info)

@strawberry.type
class Query:
    node: Node = field()
    ...

schema = strawberry.Schema(query=Query, scalar_overrides={
    GlobalID: GlobalID._get_scalar_definition(),
})
```