fix: Remove GlobalID type from relay schema.
Having this type in the schema violates the relay spec.
The GlobalID is technically an interpretation of the ID type,
and the schema should always return ID for that.

I created a new scalar type called ID. It's unclear how this
collides with the strawberry.ID, but given that strawberry.ID
isn't actually a scalar type, I am hoping it is ok.