fix(3276): Use get_template()
The `get_template()` call correctly fetches the template via the engine,
but to get the same template type as the one used in the other branch,
we need to access its template attribute, where the original template is
stored.

The engine template wrap the base template to allow direct
rendering by TemplateResponse, however this conflicts with the "template
via static files" approach taken by Strawberry.

Since I'm unfamiliar with the reasoning for this approach, we'll leave
it in tact.